+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.1.5
 
 * Change the API of `IntervalResponse.new` to accept the Rack `env` hash directly, without having the caller extract the header values manually.
+* Allow intervals to set ETags which contribute to the final ETag
 
 # 0.1.4